refactor: Merge types/docs from testing/auth/api_key_rate_limiter.pyi
- Merged type hints and docstrings from crudclient/testing/auth/api_key_rate_limiter.pyi into crudclient/testing/auth/api_key_rate_limiter.py.

- Deleted crudclient/testing/auth/api_key_rate_limiter.pyi.
